Refine fused GDN MTP configuration and tests
Rename the fused kernel selector, reduce redundant kernel cases, and cover pure, mixed, prefill, and decode model paths.
