+++ tests/kernels/mamba/test_gdn_fused_mtp.py
@@ -1,0 +1,358 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+"""Model-path tests for fused Qwen3.5 GDN MTP decode."""
+
+from __future__ import annotations
+
+import types
+from unittest.mock import patch
+
+import pytest
+import torch
+
+from vllm.platforms import current_platform
+
+if not (
+    current_platform.is_cuda() and current_platform.is_device_capability_family(100)
+):
+    pytest.skip(
+        reason="Fused GDN MTP decode requires CUDA SM10x.",
+        allow_module_level=True,
+    )
+
+from tests.v1.attention.utils import (  # noqa: E402
+    BatchSpec,
+    create_common_attn_metadata,
+    create_vllm_config,
+)
+from vllm.config import SpeculativeConfig, set_current_vllm_config  # noqa: E402
+from vllm.model_executor.layers.mamba.gdn import qwen_gdn_linear_attn  # noqa: E402
+from vllm.model_executor.layers.mamba.gdn.qwen_gdn_linear_attn import (  # noqa: E402
+    ChunkGatedDeltaRule,
+    QwenGatedDeltaNetAttention,
+)
+from vllm.model_executor.layers.mamba.mamba_utils import (  # noqa: E402
+    MambaStateShapeCalculator,
+)
+from vllm.third_party.flash_linear_attention.ops.layernorm_guard import (  # noqa: E402
+    rmsnorm_fn,
+)
+from vllm.utils.torch_utils import _encode_layer_name  # noqa: E402
+from vllm.v1.attention.backends.gdn_attn import (  # noqa: E402
+    GDNAttentionMetadataBuilder,
+)
+from vllm.v1.kv_cache_interface import MambaSpec  # noqa: E402
+
+NUM_SPEC = 3
+SPEC_TOKENS = NUM_SPEC + 1
+H = 1
+HV = 8
+K = 128
+V = 128
+CONV_KERNEL = 4
+CONV_DIM = 2 * H * K + HV * V
+BLOCK_SIZE = 16
+PREFIX = "model.layers.0.linear_attn"
+EPS = 1e-6
+
+
+class _TestGatedNorm:
+    def __init__(self, weight: torch.Tensor) -> None:
+        self.weight = weight
+        self.bias = None
+        self.eps = EPS
+        self.group_size = None
+        self.norm_before_gate = True
+        self.activation = "silu"
+
+    def __call__(self, x: torch.Tensor, z: torch.Tensor) -> torch.Tensor:
+        return rmsnorm_fn(
+            x,
+            self.weight,
+            None,
+            z=z,
+            eps=self.eps,
+            norm_before_gate=True,
+            activation=self.activation,
+        )
+
+
+def _make_vllm_config():
+    config = create_vllm_config(
+        model_name="Qwen/Qwen3.5-0.8B",
+        block_size=BLOCK_SIZE,
+        hf_config_override={"linear_key_head_dim": K},
+    )
+    config.additional_config = {"gdn_prefill_backend": "cutedsl"}
+    config.cache_config.mamba_cache_mode = "none"
+    config.speculative_config = SpeculativeConfig(
+        method="ngram", num_speculative_tokens=NUM_SPEC
+    )
+    return config
+
+
+def _build_layer(
+    vllm_config,
+    conv_state: torch.Tensor,
+    ssm_state: torch.Tensor,
+    a_log: torch.Tensor,
+    dt_bias: torch.Tensor,
+    conv_weight: torch.Tensor,
+    norm_weight: torch.Tensor,
+):
+    layer = types.SimpleNamespace(
+        prefix=PREFIX,
+        enable_packed_recurrent_decode=False,
+        disable_tp_for_ba_proj=False,
+        tp_size=1,
+        num_k_heads=H,
+        num_v_heads=HV,
+        head_k_dim=K,
+        head_v_dim=V,
+        key_dim=K,
+        value_dim=HV * V,
+        activation="silu",
+        A_log=a_log,
+        dt_bias=dt_bias,
+        conv1d=types.SimpleNamespace(weight=conv_weight, bias=None),
+        kv_cache=(conv_state, ssm_state),
+        norm=_TestGatedNorm(norm_weight),
+        layer_norm_epsilon=EPS,
+        gdn_decode_kernel="fused",
+    )
+    with set_current_vllm_config(vllm_config):
+        layer.chunk_gated_delta_rule = ChunkGatedDeltaRule()
+    for name in (
+        "rearrange_mixed_qkv",
+        "_forward_core",
+        "_forward_core_decode_spec_post_conv_fused_norm",
+        "_forward_core_decode_spec_fused_norm",
+        "_can_use_fused_gdn_mtp_decode",
+        "_rms_norm_gated_cuda",
+        "_forward_core_fused_norm",
+        "_forward_core_fused_norm_packed",
+        "split_ba",
+    ):
+        setattr(
+            layer,
+            name,
+            types.MethodType(getattr(QwenGatedDeltaNetAttention, name), layer),
+        )
+    return layer
+
+
+@torch.inference_mode()
+def test_fused_forward_uses_packed_entrypoint() -> None:
+    """Fused mode keeps projected QKVZ and BA packed through the model op."""
+    device = torch.device("cuda")
+    num_tokens = 3
+    hidden_states = torch.empty(num_tokens, 1, dtype=torch.bfloat16, device=device)
+    mixed_qkvz = torch.randn(
+        num_tokens, CONV_DIM + HV * V, dtype=torch.bfloat16, device=device
+    )
+    ba = torch.randn(num_tokens, 2 * HV, dtype=torch.bfloat16, device=device)
+    layer = types.SimpleNamespace(
+        prefix=PREFIX,
+        enable_fused_gdn_decode=True,
+        norm=types.SimpleNamespace(
+            weight=torch.empty(V, dtype=torch.bfloat16, device=device)
+        ),
+        num_v_heads=HV,
+        tp_size=1,
+        head_v_dim=V,
+        in_proj_qkvz=lambda _: (mixed_qkvz, None),
+        in_proj_ba=lambda _: (ba, None),
+        _output_projection_normalized=lambda output: output,
+    )
+    layer.forward_cuda = types.MethodType(
+        QwenGatedDeltaNetAttention.forward_cuda, layer
+    )
+
+    def packed_op(
+        actual_qkvz: torch.Tensor,
+        actual_ba: torch.Tensor,
+        output: torch.Tensor,
+        *,
+        layer_name: str,
+    ) -> None:
+        assert actual_qkvz is mixed_qkvz
+        assert actual_ba is ba
+        assert layer_name == _encode_layer_name(PREFIX)
+        output.fill_(1)
+
+    with (
+        patch.object(
+            torch.ops.vllm,
+            "qwen_gdn_attention_core_fused_norm_packed",
+            side_effect=packed_op,
+        ) as packed_mock,
+        patch.object(torch.ops.vllm, "qwen_gdn_attention_core") as triton_mock,
+    ):
+        output = layer.forward_cuda(hidden_states)
+
+    packed_mock.assert_called_once()
+    triton_mock.assert_not_called()
+    torch.testing.assert_close(output, torch.ones_like(output))
+
+
+@pytest.mark.parametrize(
+    "seq_lens,query_lens,draft_tokens,expected_fused_calls,expected_prefix_len",
+    [
+        pytest.param([128], [SPEC_TOKENS], [NUM_SPEC], 1, SPEC_TOKENS, id="pure-mtp"),
+        pytest.param(
+            [128, 96],
+            [SPEC_TOKENS, 64],
+            [NUM_SPEC, -1],
+            1,
+            SPEC_TOKENS,
+            id="mixed-mtp-first",
+        ),
+        pytest.param(
+            [96, 128],
+            [64, SPEC_TOKENS],
+            [-1, NUM_SPEC],
+            1,
+            None,
+            id="mixed-mtp-last",
+        ),
+        pytest.param([96], [64], [-1], 0, None, id="pure-prefill"),
+        pytest.param([128], [1], [-1], 0, None, id="pure-decode"),
+    ],
+)
+@torch.inference_mode()
+def test_fused_model_path_matches_reference(
+    seq_lens: list[int],
+    query_lens: list[int],
+    draft_tokens: list[int],
+    expected_fused_calls: int,
+    expected_prefix_len: int | None,
+) -> None:
+    """Fused MTP and its prefill/decode fallbacks match the reference path."""
+    torch.manual_seed(1)
+    device = torch.device("cuda")
+    vllm_config = _make_vllm_config()
+    builder = GDNAttentionMetadataBuilder(
+        kv_cache_spec=MambaSpec(
+            block_size=BLOCK_SIZE,
+            shapes=((16, 64),),
+            dtypes=(torch.float16,),
+            num_speculative_blocks=NUM_SPEC,
+        ),
+        layer_names=[PREFIX],
+        vllm_config=vllm_config,
+        device=device,
+    )
+    batch = BatchSpec(seq_lens=seq_lens, query_lens=query_lens)
+    common = create_common_attn_metadata(
+        batch, BLOCK_SIZE, device, arange_block_indices=True
+    )
+    common.block_table_tensor.add_(1)
+    with set_current_vllm_config(vllm_config):
+        metadata = builder.build(
+            common_prefix_len=0,
+            common_attn_metadata=common,
+            num_accepted_tokens=torch.ones(
+                batch.batch_size, dtype=torch.int32, device=device
+            ),
+            num_decode_draft_tokens_cpu=torch.tensor(draft_tokens, dtype=torch.int32),
+        )
+
+    assert metadata.spec_token_prefix_len == expected_prefix_len
+    state_indices = [
+        indices
+        for indices in (
+            metadata.spec_state_indices_tensor,
+            metadata.non_spec_state_indices_tensor,
+        )
+        if indices is not None and indices.numel() > 0
+    ]
+    pool_size = max(int(indices.max().item()) for indices in state_indices) + 1
+    conv_state_shape, temporal_state_shape = (
+        MambaStateShapeCalculator.gated_delta_net_state_shape(
+            1, H, HV, K, V, CONV_KERNEL, NUM_SPEC
+        )
+    )
+    conv_state_seed = 0.05 * torch.randn(
+        pool_size, *conv_state_shape, dtype=torch.bfloat16, device=device
+    )
+    ssm_state_seed = 0.01 * torch.randn(
+        pool_size, *temporal_state_shape, dtype=torch.float32, device=device
+    )
+    a_log = 0.1 * torch.randn(HV, dtype=torch.float32, device=device)
+    dt_bias = 0.1 * torch.randn(HV, dtype=torch.float32, device=device)
+    conv_weight = 0.1 * torch.randn(
+        CONV_DIM, 1, CONV_KERNEL, dtype=torch.bfloat16, device=device
+    )
+    norm_weight = torch.randn(V, dtype=torch.float32, device=device)
+    num_tokens = batch.compute_num_tokens()
+    mixed_qkv = 0.1 * torch.randn(
+        num_tokens, CONV_DIM, dtype=torch.bfloat16, device=device
+    )
+    b = 0.1 * torch.randn(num_tokens, HV, dtype=torch.bfloat16, device=device)
+    a = 0.1 * torch.randn_like(b)
+    output_gate = 0.1 * torch.randn(
+        num_tokens, HV, V, dtype=torch.bfloat16, device=device
+    )
+    mixed_qkvz = torch.cat((mixed_qkv, output_gate.flatten(1)), dim=-1)
+    ba = torch.cat((b, a), dim=-1)
+    context = types.SimpleNamespace(attn_metadata={PREFIX: metadata})
+
+    reference_layer = _build_layer(
+        vllm_config,
+        conv_state_seed.clone(),
+        ssm_state_seed.clone(),
+        a_log,
+        dt_bias,
+        conv_weight,
+        norm_weight,
+    )
+    reference_out = torch.zeros_like(output_gate)
+    with patch.object(
+        qwen_gdn_linear_attn, "get_forward_context", return_value=context
+    ):
+        reference_layer._forward_core(
+            mixed_qkv=mixed_qkv.clone(),
+            b=b,
+            a=a,
+            core_attn_out=reference_out,
+        )
+    reference_out = reference_layer.norm(reference_out, output_gate)
+
+    fused_layer = _build_layer(
+        vllm_config,
+        conv_state_seed.clone(),
+        ssm_state_seed.clone(),
+        a_log,
+        dt_bias,
+        conv_weight,
+        norm_weight,
+    )
+    context.no_compile_layers = {PREFIX: fused_layer}
+    fused_out = torch.zeros_like(output_gate)
+    fused_op = qwen_gdn_linear_attn.ops.fused_gdn_decode_post_conv_mtp
+    with (
+        patch.object(qwen_gdn_linear_attn, "get_forward_context", return_value=context),
+        patch.object(
+            qwen_gdn_linear_attn.ops,
+            "fused_gdn_decode_post_conv_mtp",
+            wraps=fused_op,
+        ) as fused_mock,
+    ):
+        torch.ops.vllm.qwen_gdn_attention_core_fused_norm_packed(
+            mixed_qkvz.clone(),
+            ba,
+            fused_out,
+            layer_name=_encode_layer_name(PREFIX),
+        )
+
+    assert fused_mock.call_count == expected_fused_calls
+    torch.testing.assert_close(
+        fused_layer.kv_cache[0], reference_layer.kv_cache[0], atol=0, rtol=0
+    )
+    torch.testing.assert_close(fused_out, reference_out, atol=3e-2, rtol=3e-2)
+    torch.testing.assert_close(
+        fused_layer.kv_cache[1],
+        reference_layer.kv_cache[1],
+        atol=3e-2,
+        rtol=3e-2,
+    )
